requirements.txt
python-dotenv
python-linkedin-v2
